+++ src/refactoring/pages/Admin/features/ProductManagement/components/ProductEditor/components/ProductUpdateForm/ProductUpdateForm.tsx
@@ -1,7 +1,6 @@
-import React, { useState } from 'react';
-
+import { DiscountDetails } from '@/refactoring/pages/Admin/features/ProductManagement/components/ProductEditor/components/ProductUpdateForm/components/DiscountDetails';
 import { useProductForm } from '@/refactoring/pages/Admin/features/ProductManagement/components/ProductEditor/components/ProductUpdateForm/hooks/useProductForm';
-import type { Discount, Product } from '@/types';
+import type { Product } from '@/types';
 
 interface ProductUpdateFormProps {
   initProduct: Product;
@@ -11,16 +10,6 @@ interface ProductUpdateFormProps {
 
 export const ProductUpdateForm = ({ initProduct, onProductUpdate, onEditComplete }: ProductUpdateFormProps) => {
   const { editingProduct, updateName, updatePrice, updateStock, updateDiscounts } = useProductForm({ initProduct });
-  const [newDiscount, setNewDiscount] = useState<Discount>({ quantity: 0, rate: 0 });
-
-  const handleAddDiscount = () => {
-    updateDiscounts([...editingProduct.discounts, newDiscount]);
-    setNewDiscount({ quantity: 0, rate: 0 });
-  };
-
-  const handleRemoveDiscount = (index: number) => {
-    updateDiscounts(editingProduct.discounts.filter((_, i) => i !== index));
-  };
 
   const handleEditComplete = () => {
     onProductUpdate(editingProduct);
@@ -56,42 +45,8 @@ export const ProductUpdateForm = ({ initProduct, onProductUpdate, onEditComplete
           className="w-full rounded border p-2"
         />
       </div>
-      {/* 할인 정보 수정 부분 */}
-      <div>
-        <h4 className="mb-2 text-lg font-semibold">할인 정보</h4>
-        {editingProduct.discounts.map((discount, index) => (
-          <div key={index} className="mb-2 flex items-center justify-between">
-            <span>
-              {discount.quantity}개 이상 구매 시 {discount.rate * 100}% 할인
-            </span>
-            <button
-              onClick={() => handleRemoveDiscount(index)}
-              className="rounded bg-red-500 px-2 py-1 text-white hover:bg-red-600"
-            >
-              삭제
-            </button>
-          </div>
-        ))}
-        <div className="flex space-x-2">
-          <input
-            type="number"
-            placeholder="수량"
-            value={newDiscount.quantity}
-            onChange={e => setNewDiscount({ ...newDiscount, quantity: parseInt(e.target.value) })}
-            className="w-1/3 rounded border p-2"
-          />
-          <input
-            type="number"
-            placeholder="할인율 (%)"
-            value={newDiscount.rate * 100}
-            onChange={e => setNewDiscount({ ...newDiscount, rate: parseInt(e.target.value) / 100 })}
-            className="w-1/3 rounded border p-2"
-          />
-          <button onClick={handleAddDiscount} className="w-1/3 rounded bg-blue-500 p-2 text-white hover:bg-blue-600">
-            할인 추가
-          </button>
-        </div>
-      </div>
+
+      <DiscountDetails discounts={editingProduct.discounts} onDiscountsUpdate={updateDiscounts} />
 
       <button
         onClick={handleEditComplete}
